+++ SK.CRM.Application/DTOs/Quote/Quote_ProductsDto.cs
@@ -10,8 +10,8 @@
 
         public string? Description { get; set; }
 
-        public double DiscountRate { get; set; } = 0;
+        public decimal DiscountRate { get; set; } = 0;
 
-        public double TaxRate { get; set; } = 0;
+        public decimal TaxRate { get; set; } = 0;
     }
 }
