+++ SK.Inventory.Application/Features/Products/Dtos/Product_GeneralInformations.cs
@@ -31,8 +31,8 @@ namespace SK.Inventory.Application.Features.Products.Dtos
 
         public string? ImageUrl { get; set; }
 
-        public double DiscountRate { get; set; } = 0;
+        public decimal DiscountRate { get; set; } = 0;
 
-        public double TaxRate { get; set; } = 0;
+        public decimal TaxRate { get; set; } = 0;
     }
 }
